requirements.txt
absl-py==2.1.0
accelerate==0.34.2
addict==2.4.0
affine==2.4.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
albucore==0.0.17
albumentations==1.4.18
alembic==1.13.1
aliyun-python-sdk-core==2.15.2
aliyun-python-sdk-kms==2.16.5
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
asttokens==3.0.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
av==12.0.0
backcall==0.2.0
banal==1.0.6
beautifulsoup4==4.12.3
bidict==0.23.1
bitsandbytes==0.42.0
bleach==6.2.0
blessed==1.20.0
blinker==1.8.2
boto3==1.39.0
botocore==1.39.0
braceexpand==0.1.7
bypy==1.8.5
certifi==2022.12.7
cffi==1.17.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
controlnet-aux==0.0.3
crcmod==1.7
cryptography==43.0.1
cycler==0.12.1
datasets==2.18.0
decorator==5.1.1
decord==0.6.0
defusedxml==0.7.1
diffusers==0.27.2
dill==0.3.8
docker-pycreds==0.4.0
docopt==0.6.2
docstring_parser==0.16
easydict==1.13
efficientnet_pytorch==0.7.1
einops==0.7.0
empatches==0.2.3
eval_type_backport==0.2.0
exceptiongroup==1.2.2
executing==2.1.0
fairscale==0.4.4
fastapi==0.108.0
fastjsonschema==2.21.1
ffmpy==0.4.0
filelock==3.14.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.3.7
flaskwebgui==1.1.4
flatbuffers==24.3.25
fonttools==4.49.0
FrEIA==0.2
frozenlist==1.4.1
fsspec==2023.5.0
ftfy==6.2.3
fvcore==0.1.5.post20221221
gast==0.6.0
gensim==4.3.3
gevent==24.2.1
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
googletrans==2.4.0
gpustat==1.1.1
grad-cam==1.5.2
gradio==4.21.0
gradio_client==0.12.0
graphviz==0.20.3
greenlet==3.0.3
grpcio==1.66.1
h11==0.9.0
h2==3.2.0
h5py==3.12.1
hpack==3.0.0
hstspreload==2024.11.1
httpcore==0.9.1
httpx==0.25.1
huggingface-hub==0.29.3
humanfriendly==10.0
hydra-core==1.3.2
hyperframe==5.2.0
idna==2.10
imageio==2.34.0
imgaug==0.4.0
importlib-resources==6.1.1
importlib_metadata==8.4.0
intel-cmplr-lib-ur==2024.2.1
intel-openmp==2024.2.1
IOPaint==1.4.3
iopath==0.1.10
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
jsonargparse==4.28.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==3.5.0
kiwisolver==1.4.5
kornia==0.6.9
lama-cleaner @ git+https://github.com/yizhangliu/lama-cleaner.git@658ee0d4257499528d8781440787ca0596794eaf
lazy_loader==0.3
libclang==18.1.1
lightning-utilities==0.11.0
linkify-it-py==2.0.3
llvmlite==0.42.0
loguru==0.7.2
Mako==1.3.2
Markdown==3.7
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.8.3
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.3
mdurl==0.1.2
MedCLIP==0.0.3
mistune==3.0.2
mkl==2024.2.1
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
ml-dtypes==0.4.1
mmengine==0.10.5
model-index==0.1.11
mpmath==1.3.0
msgpack==1.1.0
msgpack-numpy==0.4.8
multidict==6.0.5
multimethod==2.0
multiprocess==0.70.16
namex==0.0.8
narwhals==1.8.1
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.2.1
ninja==1.11.1.1
nltk==3.9.1
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvcc-cu12==12.3.107
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.16.3
onnxsim==0.4.36
open_clip_torch==2.26.1
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.1
opt_einsum==3.4.0
optree==0.12.1
ordered-set==4.1.0
orjson==3.10.7
oss2==2.17.0
packaging==24.1
pandas==1.3.5
pandocfilters==1.5.1
parso==0.8.4
peft==0.7.1
pexpect==4.9.0
pickleshare==0.7.5
piexif==1.1.3
Pillow==9.5.0
pipreqs==0.5.0
platformdirs==4.2.0
portalocker==2.10.1
prefetch_generator==1.0.3
prettytable==3.16.0
progressbar2==4.5.0
prompt_toolkit==3.0.48
protobuf==4.25.5
psutil==6.0.0
ptflops==0.7.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyarmor==9.0.6
pyarmor.cli.core==7.6.2
pyarrow==15.0.0
pyarrow-hotfix==0.6
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.11.7
pydantic_core==2.33.2
pydensecrf @ git+https://github.com/lucasb-eyer/pydensecrf.git@2723c7fa4f2ead16ae1ce3d8afe977724bb8f87f
pydub==0.25.1
pyecharts==2.0.8
Pygments==2.17.2
pynndescent==0.5.11
pynvml==11.5.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.7.2
python-utils==3.8.2
pytorch-lightning==1.9.5
pytz==2023.4
PyYAML==6.0.1
pyzmq==26.2.0
qudida==0.0.4
qwen-vl-utils==0.0.10
rasterio==1.4.2
referencing==0.35.1
regex==2023.12.25
requests==2.28.2
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.4.2
rpds-py==0.20.0
ruff==0.6.3
s3transfer==0.13.0
safetensors==0.4.5
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.2.0
sentry-sdk==2.14.0
setproctitle==1.3.3
shapely==2.0.4
shellingham==1.5.4
simple-websocket==1.0.0
simplejson==3.20.1
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==1.4.52
stack-data==0.6.3
starlette==0.32.0.post1
submitit==1.5.2
supervision==0.3.2
swankit==0.2.4
swanlab==0.6.4
sympy==1.12
tabulate==0.9.0
tbb==2021.13.1
tencentcloud-sdk-python==3.0.1241
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
tensorpack==0.11
termcolor==2.4.0
textaugment==2.0.0
textblob==0.18.0.post0
thop==0.1.1.post2209072238
threadpoolctl==3.3.0
tifffile==2024.2.12
timm==0.9.12
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
torch==2.2.0
torch-tb-profiler==0.4.3
torchaudio==2.2.0
torchmetrics==0.10.3
torchpack==0.3.1
torchsummary==1.5.1
torchvision==0.17.0
tornado==6.4.2
tqdm==4.65.2
traitlets==5.14.3
transformers==4.49.0
triton==2.2.0
ttach==0.0.3
typer==0.12.5
typer-config==1.4.0
typeshed_client==2.5.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
umap-learn==0.5.5
urllib3==1.26.20
uvicorn==0.30.6
visual-genome==1.1.1
wandb==0.18.1
wcwidth==0.2.13
webdataset==0.2.100
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.4
wget==3.2
wldhx.yadisk-direct==0.0.6
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.2
yarg==0.1.9
yarl==1.9.4
zipp==3.20.1
zope.event==5.0
zope.interface==7.0.3
